edit catalog to pydantic and select on iter_docs

+++ datasets/cfg.py
@@ -1,4 +1,5 @@
-from polars.datatypes import DataType, Utf8, Int32, Float32, List, Struct, Field
+from polars.datatypes import Utf8, Int64, Float32, List, Struct, Field
+
 
 class Storage:
     base_path: str = "gs://pinecone-datasets-dev"
@@ -8,10 +9,10 @@ class Schema:
     documents = {
         "id": Utf8,
         "values": List(Float32),
-        "sparse_values": Struct([Field('indices', List(Int32)), Field('values', List(Float32))])
+        "sparse_values": Struct([Field('indices', List(Int64)), Field('values', List(Float32))])
     }
     queries = {
         "id": Utf8,
         "values": List(Float32),
-        "sparse_values": Struct([Field('indices', List(Int32)), Field('values', List(Float32))])
+        "sparse_values": Struct([Field('indices', List(Int64)), Field('values', List(Float32))])
     }
